==> src/1_additional_data.ipynb <==
# Load List of Kings and Queens
This Scripts generates a list with all members of the Kantonsrat back to the age of the dinosaurs. It states as well who was when president of the Kantonsrat.  

Load from xlsx because:
* CSVs are corrupt
* API from Parlamentsdienste is a pain  
* Be aware: The data is messy. Duplicated entries, typos, just not as it should be...

Download at the bottom: https://www.zh.ch/de/politik-staat/wahlen-abstimmungen/kantons-regierungsratswahlen/mitglieder-kantonsrats-ab-1803.html#-373901206

import pandas as pd
from datetime import datetime
from pathlib import Path
import numpy as np
import json

df_person = pd.read_excel(Path('../data/kantonsrathistory.xlsx'), 'Personen')
df_funktion = pd.read_excel(Path('../data/kantonsrathistory.xlsx'), 'Funktionen')
df_partei = pd.read_excel(Path('../data/kantonsrathistory.xlsx'), 'Parteien')
df_einsitz = pd.read_excel(Path('../data/kantonsrathistory.xlsx'), 'Einsitze')

df_person = df_person[df_person.DATUM_GEBURT_JAHR.notna()]
df_person = df_person[df_person.DATUM_GEBURT_JAHR> 1900]
df_einsitz = df_einsitz[df_einsitz.DATUM_EINTRITT_JAHR.notna()]
df_einsitz = df_einsitz[df_einsitz.RAT == 'Kantonsrat']
df_funktion = df_funktion[df_funktion.DATUM_VON_JAHR.notna()]
df_funktion = df_funktion[df_funktion.FUNKTIONSBEZEICHNUNG.notna()]
df_partei = df_partei[df_partei.PARTEIBEZEICHNUNG.notna()]

records = []

for i, row in df_person.iterrows():

    # Search if name already in list. Yes, there are duplicates in the data... :(
    filtered = list(filter(lambda r: (
        (r['name'] == row['NACHNAME'].strip()) and
        (r['vorname'] == row['VORNAME'].strip()) and
        (r['jahrgang'] == row['DATUM_GEBURT_JAHR'])
    ), records))
    if len(filtered) > 0:
        print("Duplicate found. Add to original: %s %s" % (row['NACHNAME'], row['VORNAME'].strip()))
        #print(filtered)

        record = filtered[0]

    else:
        # No duplicate found, add it

        record = {
            'id': row['ID_PERSON_NEW'],
            'name': row['NACHNAME'].strip(),
            'vorname': row['VORNAME'].strip(),
            'geschlecht': row['GESCHLECHT'].strip(),
            'jahrgang': row['DATUM_GEBURT_JAHR'],
            'einsitz': [],
            'partei': [],
            'funktion': []
        }

        records.append(record)

    # Add Einsitz
    df = df_einsitz[df_einsitz.ID_PERSON_NEW == row['ID_PERSON_NEW']]
    for y, row2 in df.iterrows():

        if np.isnan(row2['DATUM_AUSTRITT_JAHR']):
            end = datetime(2100, 12, 31)
        else:
            end = datetime(int(row2['DATUM_AUSTRITT_JAHR']), int(row2['DATUM_AUSTRITT_MONAT']), int(row2['DATUM_AUSTRITT_TAG']))

        record['einsitz'].append({
            'start': datetime(int(row2['DATUM_EINTRITT_JAHR']), int(row2['DATUM_EINTRITT_MONAT']), int(row2['DATUM_EINTRITT_TAG'])),
            'end': end
        })

    # Add Party
    df = df_partei[df_partei.ID_PERSON_NEW == row['ID_PERSON_NEW']]
    try:
        for y, row2 in df.iterrows():
            if not(np.isnan(row2['DATUM_BIS_MONAT'])) and not(np.isnan(row2['DATUM_BIS_JAHR'])):
                end = datetime(int(row2['DATUM_BIS_JAHR']), int(row2['DATUM_BIS_MONAT']), int(row2['DATUM_BIS_TAG']))
            else:
                end = datetime(2100, 12, 31)

            if np.isnan(row2['DATUM_VON_JAHR']):
                start = record['einsitz'][0]['start']
            else:
                start = datetime(int(row2['DATUM_VON_JAHR']), int(row2['DATUM_VON_MONAT']), int(row2['DATUM_VON_TAG']))

            record['partei'].append({
                'bezeichnung': row2['PARTEIBEZEICHNUNG'],
                'start': start,
                'end': end
            })
        # Reorder
        record['partei'] = sorted(record['partei'], key = lambda x: x['start'])
    except:
        print(row2)
        raise
        
    # Add Funktion
    df = df_funktion[df_funktion.ID_PERSON_NEW == row['ID_PERSON_NEW']]
    try:
        for y, row2 in df.iterrows():

            # Skip if "Mitglied"
            if row2['FUNKTIONSBEZEICHNUNG'].lower() == 'mitglied': continue

            # Assume, that Amtsjahr starts at 1. Mai and ends at 30. April
            if not np.isnan(row2['DATUM_VON_JAHR']):
                day = 1 if np.isnan(row2['DATUM_VON_TAG']) else int(row2['DATUM_VON_TAG'])
                month = 5 if np.isnan(row2['DATUM_VON_MONAT']) else int(row2['DATUM_VON_MONAT'])
                start = datetime(int(row2['DATUM_VON_JAHR']), month, day)
            else:
                start = datetime(2100, 12, 31)
                print("Hmmm")


            if np.isnan(row2['DATUM_BIS_JAHR']):
                end = datetime(2100, 12, 31)
            else:
                day = 30 if np.isnan(row2['DATUM_BIS_TAG']) else int(row2['DATUM_BIS_TAG'])
                month = 4 if np.isnan(row2['DATUM_BIS_MONAT']) else int(row2['DATUM_BIS_MONAT'])
                end = datetime(int(row2['DATUM_BIS_JAHR']), month, day)                

            record['funktion'].append({
                'bezeichnung': row2['FUNKTIONSBEZEICHNUNG'],
                'start': start,
                'end': end
            })

    except:
        print(row2)
        raise

# Drop by ID (duplicated entries)

# Not needed in new file version
"""
print(len(records))
remove = None
for r in records:
    if r['id'] == 21252:
        remove = r
        break

if remove:
    records.remove(remove)

print(len(records))
"""

# Fix Christoph Fischbach
r = list(filter(lambda x: x['id'] == 21362, records))[0]
r['einsitz'].append({
    "start": "2021-11-08 00:00:00",
    "end": "2100-12-31 00:00:00"
})

r['partei'].append({
    "bezeichnung": "SP",
    "start": "2021-11-08 00:00:00",
    "end": "2100-12-31 00:00:00"
})

# Fix Cristina Cortellini
r = list(filter(lambda x: x['id'] == 21445, records))[0]
r['name'] = 'Wyss-Cortellini'

# Fix Hansruedi Schmid
r = list(filter(lambda x: (x['vorname'] == 'Hansruedi') and (x['name'] == 'Schmid'), records))[0]
r['partei'][0]['end'] = datetime(2005, 7, 11)

# Fix Jean-Jacques Bertschi
r = list(filter(lambda x: (x['vorname'] == 'Jean-Jacques') and (x['name'] == 'Bertschi'), records))[0]
r['partei'][0]['start'] = datetime(1992, 1, 2)

# Fix Chanson Robert
r = list(filter(lambda x: (x['vorname'] == 'Robert') and (x['name'] == 'Chanson'), records))[0]
r['partei'][0]['end'] = datetime(2002, 10, 21)

# Fix Stefanie Huber
r = list(filter(lambda x: x['id'] == 20494, records))
records.remove(r[0])

r = list(filter(lambda x: x['id'] == 18066, records))
r[0]['einsitz'].append({
        "start": datetime(2012, 2, 13),
        "end": "2015-05-17 00:00:00"
      })
r[0]['partei'][0]['start'] = '2012-02-13 00:00:00'
r[0]['einsitz'].sort(key=lambda x: x['start'])

# Add manually
# Not needed anymore
"""
records.append({
    "id": 0,
    "name": "Hoss-Blatter",
    "vorname": "Corinne",
    "geschlecht": "w",
    "jahrgang": 1937.0,
    "einsitz": [{"start": "2020-02-24 00:00:00", "end": "2100-12-31 00:00:00"}],
    "partei": [{"bezeichnung": "FDP", "start": "2020-02-24 00:00:00", "end": "2100-12-31 00:00:00"}],
    "funktion": []})
"""

with open(Path('../export/mitglieder.json'), 'w', encoding='utf-8') as f:
    json.dump(records, f, default=str, ensure_ascii=False, indent=2)
