==> src/99_figures.ipynb <==
# Some numbers

import pandas as pd
from pathlib import Path
import glob
from pdfminer.pdfparser import PDFParser
from pdfminer.pdfdocument import PDFDocument
from pdfminer.pdfinterp import resolve1

## Count Protokolle and Pages

docs = 0
pages = 0

for f in glob.glob(str(Path('./../export/Files/*.pdf'))):

    parser = PDFParser(open(f, 'rb'))
    document = PDFDocument(parser)
    docs += 1
    pages += resolve1(document.catalog['Pages'])['Count']

print("Dokumente: %s" % docs)
print("Seiten: %s" % pages)

df = pd.concat([
    pd.read_csv(Path('../export/votum/votum_0.csv')),
    pd.read_csv(Path('../export/votum/votum_1.csv'))
])

print("Voten: %s" % len(df))
print("First: %s" % df['sitzung_date'].min())
print("Last: %s" % df['sitzung_date'].max())

## Count Geschäfte

df = pd.read_csv(Path('../export/geschaefte.csv'), encoding='UTF-8')

print("Geschäfte: %s" % len(df))
print("First: %s" % df['start'].min())
print("Last: %s" % df['start'].max())

## Alle Sitzungen

df_votum_raw = pd.concat([
    pd.read_csv(Path('../export/votum/votum_0.csv')),
    pd.read_csv(Path('../export/votum/votum_1.csv'))
])

# Remove non members (mostly former members who are now in the Regierungsrat)
df_votum_raw = df_votum_raw[df_votum_raw.ismember == True]

# Typecast
df_votum_raw['sitzung_date'] = pd.to_datetime(df_votum_raw['sitzung_date'])

# Remove empty texts
df_votum_raw = df_votum_raw[df_votum_raw.text.notna()]

# Replace CVP with Die Mitte
df_votum_raw.loc[df_votum_raw.partei.str.lower() == 'cvp', 'partei'] = "Die Mitte"

df_votum_raw['g'] = df_votum_raw['geschlecht']
df_votum_raw.loc[df_votum_raw.funktion.notna(), 'g'] = 'präsidium'
df_votum_raw.loc[df_votum_raw.ismember == False, 'g'] = 'nomember'

len(df_votum_raw)

df = df_votum_raw
df = df[df.funktion.isna()]
df = df.groupby('g').count()
df['%'] = 100 / df['name'].sum() * df['name']

df

## Since 2019

df = df_votum_raw[df_votum_raw.sitzung_start >= '2019-05-01']
#df = df[df.funktion.isna()]
df = df.groupby('geschlecht').count()
df

## Sitzung 2021-08-23

df = df_votum_raw[df_votum_raw.sitzung_date == '2021-08-23'].copy()

df = df[df.funktion.isna()]

df.groupby('g').count()

100 / (41+14) * 41

df[(df.page >= 49) & (df.page <= 65)].groupby('geschlecht').count()
